after sign in redirect to first org by default

+++ app/controllers/concerns/authentication.rb
@@ -11,7 +11,13 @@ module Authentication
   private
 
   def after_sign_in_path_for(resource)
-    if session[:new_user]
+    stored_location = stored_location_for(resource)
+    return stored_location if stored_location
+
+    if resource.organizations.any?
+      session.delete(:new_user) if session[:new_user]
+      organization_path(resource.organizations.first)
+    elsif session[:new_user]
       session.delete(:new_user)
       # You can send new users to onboarding, billing, or somewhere else
       root_path
